sigh

+++ benches/pir.rs
@@ -693,7 +693,7 @@ macro_rules! pir_decode_pung {
                 // Generate keys (desired indexes)
                 let mut key_set: HashSet<usize> = HashSet::new();
                 while key_set.len() < $k { 
-                    key_set.insert(rng.next_u32() as usize % $num);
+                    key_set.insert(rng.next_u32() as usize % $num as usize);
                 }
 
                 let keys: Vec<usize> = key_set.drain().collect();
